Move plate from v04 to common
